More refacto
